refactor: move translation tests to test_api_translations.py
